# docs/swagger.yaml
basePath: /
definitions:
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  user.UserProfileDataModel:
    properties:
      date_birth:
        type: string
      gender:
        type: boolean
      name:
        type: string
      nickname:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      surname:
        type: string
    required:
    - name
    - nickname
    - surname
    type: object
...
host: localhost:5000
info:
  contact: {}
  title: MISU Main Server
  version: "1.0"
paths:
...
  /user/profile/get:
    post:
      consumes:
      - application/json
      description: Get user profile
      operationId: get-profile
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            $ref: '#/definitions/user.UserProfileDataModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: GetProfile
      tags:
      - profile
  /user/profile/update:
    post:
      consumes:
      - application/json
      description: Update user profile
      operationId: update-profile
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user.UserProfileDataModel'
      produces:
      - application/json
      responses:
        "200":
          description: data
          schema:
            $ref: '#/definitions/user.UserProfileDataModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: UpdateProfile
      tags:
      - profile
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"